+++ mex/extractors/datenkompass/extract.py
@@ -1,11 +1,13 @@
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING
 
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.exceptions import MExError
 from mex.common.identity import get_provider
+from mex.common.models import MEX_PRIMARY_SOURCE_STABLE_TARGET_ID
 
 if TYPE_CHECKING:
-    from mex.common.models import AnyMergedModel, MergedPrimarySource
+    from mex.common.models import AnyMergedModel
+    from mex.common.types import MergedIdentifier
 
 
 def get_merged_items(
@@ -38,10 +40,51 @@ def get_merged_items(
     return list(response)
 
 
+def get_extracted_item_stable_target_ids(
+    entity_type: list[str],
+    referenced_identifier: list[str] | None,
+) -> list[MergedIdentifier]:
+    """Fetch extracted items from backend and return their stableTargetId.
+
+    Args:
+        entity_type: List of entity types.
+        referenced_identifier: list of MergedIdentifiers to filter for
+
+    Returns:
+        List of stableTargetIds of extracted items of the given entity type(s).
+    """
+    reference_field = "stableTargetId"
+    connector = BackendApiConnector.get()
+
+    response = connector.fetch_extracted_items(
+        entity_type=entity_type,
+        referenced_identifier=referenced_identifier,
+        reference_field=reference_field,
+        skip=0,
+        limit=1,
+    )
+    total_item_number = response.total
+    item_number_limit = 100  # 100 is the maximum possible number per get-request
+
+    extracted_items = []
+    for item_counter in range(0, total_item_number, item_number_limit):
+        extracted_items.extend(
+            connector.fetch_extracted_items(
+                entity_type=entity_type,
+                referenced_identifier=referenced_identifier,
+                reference_field=reference_field,
+                skip=item_counter,
+                limit=item_number_limit,
+            ).items
+        )
+
+    return [item.stableTargetId for item in extracted_items]
+
+
 def get_filtered_primary_source_ids(
     filtered_primary_sources: list[str] | str | None,
 ) -> list[str]:
-    """Get the IDs of the relevant primary sources.
+    """Get a list of MergedIdentifier of filtered primary sources.
 
     Args:
         filtered_primary_sources: List of primary sources.
@@ -56,16 +99,12 @@ def get_filtered_primary_source_ids(
     if isinstance(filtered_primary_sources, str):
         filtered_primary_sources = [filtered_primary_sources]
 
-    merged_primary_sources = cast(
-        "list[MergedPrimarySource]",
-        get_merged_items(entity_type=["MergedPrimarySource"]),
-    )
-
     provider = get_provider()
 
     return [
-        str(mps.identifier)
-        for mps in merged_primary_sources
-        if provider.fetch(stable_target_id=mps.identifier)[0].identifierInPrimarySource
-        in filtered_primary_sources
+        provider.fetch(
+            identifier_in_primary_source=fps,
+            had_primary_source=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
+        )[0].stableTargetId
+        for fps in filtered_primary_sources
     ]