+++ mex/extractors/datenkompass/filter.py
@@ -1,8 +1,10 @@
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING, Any, cast
 
+from mex.common.exceptions import MExError
 from mex.common.logging import logger
 from mex.common.organigram.helpers import find_descendants
 from mex.extractors.datenkompass.extract import (
+    get_extracted_item_stable_target_ids,
     get_merged_items,
 )
 from mex.extractors.settings import Settings
@@ -12,10 +14,12 @@ if TYPE_CHECKING:
         MergedActivity,
         MergedOrganization,
         MergedOrganizationalUnit,
+        MergedResource,
     )
     from mex.common.types import (
         MergedOrganizationalUnitIdentifier,
     )
+    from mex.extractors.datenkompass.models.mapping import DatenkompassFilterMapping
 
 
 def filter_activities_for_organization_and_unit(
@@ -70,6 +74,57 @@ def filter_activities_for_organization_and_unit(
     return filtered_merged_activities_by_unit
 
 
+def filter_merged_resources_by_unit(
+    merged_resources_by_primary_source: dict[str, list[MergedResource]],
+    resource_filter_mapping: DatenkompassFilterMapping,
+    merged_organizational_units_by_id: dict[
+        MergedOrganizationalUnitIdentifier, MergedOrganizationalUnit
+    ],
+) -> dict[str, dict[str, list[MergedResource]]]:
+    """Filter the merged resources by (unit and its childunits) in field unitInCharge.
+
+    Args:
+        merged_resources_by_primary_source: merged resources by primary source.
+        resource_filter_mapping: Datenkompass resource filter mapping
+        merged_organizational_units_by_id: all merged units by their id
+
+    Returns:
+        filtered list of merged resources by primary source by unit.
+    """
+    allowedunits_by_filterunits = {
+        filter_unit: set(
+            find_descendant_units(merged_organizational_units_by_id, filter_unit)
+        )
+        for filter_unit in (
+            resource_filter_mapping.fields[1].filterRules[0].forValues or []
+        )
+    }
+    result_resources_by_primary_source_by_unit: dict[
+        str, dict[str, list[MergedResource]]
+    ] = {}
+
+    for filter_unit, allowed_units in allowedunits_by_filterunits.items():
+        result_resources_by_primary_source: dict[str, list[MergedResource]] = {}
+        for (
+            primary_source,
+            merged_resources,
+        ) in merged_resources_by_primary_source.items():
+            allowed_merged_resources = [
+                item
+                for item in merged_resources
+                if allowed_units.intersection(item.unitInCharge)
+            ]
+            if allowed_merged_resources:
+                result_resources_by_primary_source[primary_source] = (
+                    allowed_merged_resources
+                )
+        result_resources_by_primary_source_by_unit[filter_unit] = (
+            result_resources_by_primary_source
+        )
+
+    return result_resources_by_primary_source_by_unit
+
+
 def find_descendant_units(
     merged_organizational_units_by_id: dict[
         MergedOrganizationalUnitIdentifier, MergedOrganizationalUnit
@@ -102,3 +157,53 @@ def find_descendant_units(
     descendants.append(parent_id)
 
     return descendants
+
+
+def filter_merged_items_for_primary_source(
+    merged_items_by_primary_source: dict[str, list[MergedResource]],
+    entity_type: str,
+) -> dict[str, list[MergedResource]]:
+    """Filter the merged items for primary source as defined in settings.
+
+     Special treatment for items which were created/edited in editor: filter those
+     merged items out, which are referenced via stableTargetID by an extracted item,
+     to keep only those merged items which consist only of rules
+
+    Args:
+        merged_items_by_primary_source: merged items dictionary by primary source.
+        entity_type: entity type to of merged items
+
+    Settings: primary source which needs to be filtered
+
+    Returns:
+        dictionary with list of filtered merged items
+    """
+    primary_source_filter = "mex-editor"
+    extracted_item_stid: set[Any] = set()
+    stepwidth = 100
+
+    if primary_source_filter not in merged_items_by_primary_source:
+        msg = "No mex-editor items in dictionary"
+        raise MExError(msg)
+
+    mex_editor_merged_items = merged_items_by_primary_source[primary_source_filter]
+    mex_editor_merged_item_ids = [
+        str(item.identifier) for item in mex_editor_merged_items
+    ]
+    for counter in range(0, len(mex_editor_merged_item_ids), stepwidth):
+        extracted_item_stid.update(
+            set(
+                get_extracted_item_stable_target_ids(
+                    [entity_type],
+                    mex_editor_merged_item_ids[counter : counter + stepwidth],
+                )
+            )
+        )
+    new_merged_items_by_primary_source = dict(merged_items_by_primary_source)
+    new_merged_items_by_primary_source[primary_source_filter] = [
+        item
+        for item in mex_editor_merged_items
+        if item.identifier not in extracted_item_stid
+    ]
+
+    return new_merged_items_by_primary_source