+++ mex/extractors/datenkompass/models/item.py
@@ -87,6 +87,7 @@ class DatenkompassResource(BaseModel):
     frequenz: str | None = Field(None, alias="Frequenz der Aktualisierung")
     kontakt: str | None = Field(None, alias="Kontakt (Herausgeber)")
     organisationseinheit: str = Field(..., alias="Organisationseinheit")
+    startdatum: str | None = Field(None, alias="Startdatum")
     titel: str = Field(..., alias="Titel")
     schlagwort: str = Field(..., alias="Schlagwort")
     voraussetzungen: str = Field(